move tests to the version-specific directory
